Create Day 14
Ambitious Kid
Chaneka, Pak Chanek's child, is an ambitious kid, so Pak Chanek gives her the following problem to test her ambition.
Given an array of integers [A1, A2, A3,…, AN]. In one operation, Chaneka can choose one element, and and then increase or decrease the element's value by 1. Chaneka can do that operation multiple times, even for different elements. What is the minimum number of operations that must be done to make it such that A1×A2×A3×…×AN=0?