Create Day 16
Slavic is preparing a present for a friend's birthday. He has an array of n digits and the present will be the product of all these digits. Because Slavic is a good kid who wants to make the biggest product possible, he wants to add 1 to exactly one of his digits. What is the maximum product Slavic can make?