Create Day 20
Maximum Increase
You are given an array consisting of n integers. Your task is to find the maximum length of an increasing subarray of the given array. A subarray is the sequence of consecutive elements of the array. Subarray is called increasing if each element of this subarray is strictly greater than the previous.