Create Day 22
Numbers 1,2,3,…n (each integer from 1 to n once) are written on a board. In one operation you can erase any two numbers a and b from the board and write one integer a+b2 rounded up instead. You should perform the given operation n−1 times and make the resulting number that will be left on the board as small as possible. For example, if n=4, the following course of action is optimal: choose a=4 and b=2, so the new number is 3, and the whiteboard contains [1,3,3];choose a=3 and b=3, so the new number is 3, and the whiteboard contains [1,3];choose a=1 and b=3, so the new number is 2, and the whiteboard contains [2].It'seasy to see that after n−1 operations, there will be left only one number. Your goal is to minimize it.